Refactor circuit simulator
