+++ src/qutip_qip/circuit/simulator/circuitsimulator.py
@@ -9,6 +9,7 @@ from qutip_qip.operations import (
     expand_operator,
 )
 from qutip import ket2dm, Qobj, tensor
+from qutip_qip.circuit.simulator import CircuitResult
 import warnings
 
 
@@ -182,7 +183,7 @@ def gate_sequence_product(U_list, ind_list):
         # expand the tensor_list to a unitary and call gate_sequence_product
         # recursively on the rest of the U_list.
         if len(overall_inds) == 1 and len(overall_inds[0]) == num_qubits:
-            # TODO undefined variable tensor_list
+            # FIXME undefined variable tensor_list
             U_overall, overall_inds = _expand_overall(
                 tensor_list, overall_inds
             )
@@ -581,99 +582,3 @@ class CircuitSimulator:
                 "mode {} is not available.".format(self.mode)
             )
         return state
-
-
-class CircuitResult:
-    """
-    Result of a quantum circuit simulation.
-    """
-
-    def __init__(self, final_states, probabilities, cbits=None):
-        """
-        Store result of CircuitSimulator.
-
-        Parameters
-        ----------
-        final_states: list of Qobj.
-            List of output kets or density matrices.
-
-        probabilities: list of float.
-            List of probabilities of obtaining each output state.
-
-        cbits: list of list of int, optional
-            List of cbits for each output.
-        """
-
-        if isinstance(final_states, Qobj) or final_states is None:
-            self.final_states = [final_states]
-            self.probabilities = [probabilities]
-            if cbits:
-                self.cbits = [cbits]
-        else:
-            inds = list(
-                filter(
-                    lambda x: final_states[x] is not None,
-                    range(len(final_states)),
-                )
-            )
-            self.final_states = [final_states[i] for i in inds]
-            self.probabilities = [probabilities[i] for i in inds]
-            if cbits:
-                self.cbits = [cbits[i] for i in inds]
-
-    def get_final_states(self, index=None):
-        """
-        Return list of output states.
-
-        Parameters
-        ----------
-        index: int
-            Indicates i-th state to be returned.
-
-        Returns
-        -------
-        final_states: Qobj or list of Qobj.
-            List of output kets or density matrices.
-        """
-
-        if index is not None:
-            return self.final_states[index]
-        return self.final_states
-
-    def get_probabilities(self, index=None):
-        """
-        Return list of probabilities corresponding to the output states.
-
-        Parameters
-        ----------
-        index: int
-            Indicates i-th probability to be returned.
-
-        Returns
-        -------
-        probabilities: float or list of float
-            Probabilities associated with each output state.
-        """
-
-        if index is not None:
-            return self.probabilities[index]
-        return self.probabilities
-
-    def get_cbits(self, index=None):
-        """
-        Return list of classical bit outputs corresponding to the results.
-
-        Parameters
-        ----------
-        index: int
-            Indicates i-th output, probability pair to be returned.
-
-        Returns
-        -------
-        cbits: list of int or list of list of int
-            list of classical bit outputs
-        """
-
-        if index is not None:
-            return self.cbits[index]
-        return self.cbits
