+++ src/qutip_qip/circuit/simulator/result.py
@@ -1,0 +1,96 @@
+from qutip import Qobj
+
+class CircuitResult:
+    """
+    Result of a quantum circuit simulation.
+    """
+
+    def __init__(self, final_states, probabilities, cbits=None):
+        """
+        Store result of CircuitSimulator.
+
+        Parameters
+        ----------
+        final_states: list of Qobj.
+            List of output kets or density matrices.
+
+        probabilities: list of float.
+            List of probabilities of obtaining each output state.
+
+        cbits: list of list of int, optional
+            List of cbits for each output.
+        """
+
+        if isinstance(final_states, Qobj) or final_states is None:
+            self.final_states = [final_states]
+            self.probabilities = [probabilities]
+            if cbits:
+                self.cbits = [cbits]
+        else:
+            inds = list(
+                filter(
+                    lambda x: final_states[x] is not None,
+                    range(len(final_states)),
+                )
+            )
+            self.final_states = [final_states[i] for i in inds]
+            self.probabilities = [probabilities[i] for i in inds]
+            if cbits:
+                self.cbits = [cbits[i] for i in inds]
+
+    def get_final_states(self, index=None):
+        """
+        Return list of output states.
+
+        Parameters
+        ----------
+        index: int
+            Indicates i-th state to be returned.
+
+        Returns
+        -------
+        final_states: Qobj or list of Qobj.
+            List of output kets or density matrices.
+        """
+
+        if index is not None:
+            return self.final_states[index]
+        return self.final_states
+
+    def get_probabilities(self, index=None):
+        """
+        Return list of probabilities corresponding to the output states.
+
+        Parameters
+        ----------
+        index: int
+            Indicates i-th probability to be returned.
+
+        Returns
+        -------
+        probabilities: float or list of float
+            Probabilities associated with each output state.
+        """
+
+        if index is not None:
+            return self.probabilities[index]
+        return self.probabilities
+
+    def get_cbits(self, index=None):
+        """
+        Return list of classical bit outputs corresponding to the results.
+
+        Parameters
+        ----------
+        index: int
+            Indicates i-th output, probability pair to be returned.
+
+        Returns
+        -------
+        cbits: list of int or list of list of int
+            list of classical bit outputs
+        """
+
+        if index is not None:
+            return self.cbits[index]
+        return self.cbits
